eliminating the weather.py module by placing the two functions into sub-functions of the Access class in access.py. Then fixing tau_intra definition in splan.py. Lastly, moving all functions out of dynamic.py and into plot.py, cleaning up all dn. to pl. calls throughout driver, webapp, and test_sample, and then deleting the dynamic.py module.
